@@ -57,7 +57,7 @@ Widget buildView(
                         child: Navigator(
                           initialRoute: '/',
                           onGenerateRoute: (RouteSettings settings) {
-                            WidgetBuilder builder;
+                            late WidgetBuilder builder;
                             switch (settings.name) {
                               case '/':
                                 builder = (context) {
@@ -99,9 +99,10 @@ Widget buildView(
                                                           int index) {
                                                     MapEntry locale =
                                                         locales[index];
-                                                    final String localeString =
+                                                    final String? localeString =
                                                         LocaleNames.of(context)
-                                                            .nameOf(locale.key);
+                                                            ?.nameOf(
+                                                                locale.key);
                                                     return BaseButton(
                                                       padding: EdgeInsets.zero,
                                                       child: Row(
