@@ -4,8 +4,8 @@ import 'package:fish_redux/fish_redux.dart';
 import 'package:palette_generator/palette_generator.dart';
 
 class ImageColorsState implements Cloneable<ImageColorsState> {
-  File file;
-  PaletteGenerator paletteGenerator;
+  File? file;
+  PaletteGenerator? paletteGenerator;
 
   @override
   ImageColorsState clone() {
