@@ -2,8 +2,7 @@ import 'package:fish_redux/fish_redux.dart';
 import 'package:flutter_easy_example/api/tu_chong/tu_chong_model.dart';
 
 class PhotoViewState implements Cloneable<PhotoViewState> {
-
-  TuChongModel data;
+  TuChongModel? data;
 
   @override
   PhotoViewState clone() {
@@ -13,4 +12,9 @@ class PhotoViewState implements Cloneable<PhotoViewState> {
 
 PhotoViewState initState(Map<String, dynamic> args) {
   return PhotoViewState()..data = args["data"];
+}
+
+extension TuChongModelExtension on Images {
+  String get imageURL =>
+      "https://photo.tuchong.com/" + "$userId" + "/f/" + "$imgId" + ".jpg";
 }
