@@ -2,7 +2,7 @@ import 'package:fish_redux/fish_redux.dart';
 import 'package:flutter/widgets.dart';
 
 class HomeState implements Cloneable<HomeState> {
-  AnimationController animationController;
+  AnimationController? animationController;
 
   @override
   HomeState clone() {
