@@ -1,24 +1,25 @@
-/// userId : "1"
-/// nickname : "用户flutter"
-/// avatar : ""
-
 class UserModel {
-  String userId;
-  String nickname;
-  String avatar;
+  String? userId;
+  String? nickname;
+  String? avatar;
 
-  static UserModel fromMap(Map<String, dynamic> map) {
-    if (map == null) return null;
-    UserModel modelBean = UserModel();
-    modelBean.userId = map['userId'];
-    modelBean.nickname = map['nickname'];
-    modelBean.avatar = map['avatar'];
-    return modelBean;
+  UserModel({
+      this.userId, 
+      this.nickname, 
+      this.avatar});
+
+  UserModel.fromJson(dynamic json) {
+    userId = json["userId"];
+    nickname = json["nickname"];
+    avatar = json["avatar"];
   }
 
-  Map toJson() => {
-    "userId": userId,
-    "nickname": nickname,
-    "avatar": avatar,
-  };
+  Map<String, dynamic> toJson() {
+    var map = <String, dynamic>{};
+    map["userId"] = userId;
+    map["nickname"] = nickname;
+    map["avatar"] = avatar;
+    return map;
+  }
+
 }
