@@ -53,7 +53,7 @@ class _AppState extends State<App> {
       logWTF("$locale");
       if (locale != null) {
         lastStorageLocale = "$locale";
-        await setStorageString(_localeKey, lastStorageLocale);
+        await setStorageString(_localeKey, lastStorageLocale!);
       } else {
         lastStorageLocale = null;
         removeStorage(_localeKey);
@@ -84,9 +84,11 @@ class _AppState extends State<App> {
       supportedLocales: S.delegate.supportedLocales,
       locale: lastLocale,
       localeResolutionCallback:
-          (Locale locale, Iterable<Locale> supportedLocales) {
+          (Locale? locale, Iterable<Locale>? supportedLocales) {
         logWTF("localeResolutionCallback: $locale");
-        if (lastLocale == null || !S.delegate.isSupported(locale)) {
+        if (lastLocale == null ||
+            locale == null ||
+            !S.delegate.isSupported(locale)) {
           return null;
         }
         return locale;
