@@ -7,7 +7,7 @@
 // ignore_for_file:unnecessary_brace_in_string_interps, unnecessary_new
 // ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
 // ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
-// ignore_for_file:unused_import, file_names
+// ignore_for_file:unused_import, file_names, avoid_escaping_inner_quotes
 
 import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';
@@ -20,26 +20,39 @@ class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'en';
 
   final messages = _notInlinedMessages(_notInlinedMessages);
-  static _notInlinedMessages(_) => <String, Function> {
-    "account" : MessageLookupByLibrary.simpleMessage("Account"),
-    "cancel" : MessageLookupByLibrary.simpleMessage("Cancel"),
-    "example" : MessageLookupByLibrary.simpleMessage("Example"),
-    "example_AgreeToContinueTip" : MessageLookupByLibrary.simpleMessage("Need to agree to continue using"),
-    "example_And" : MessageLookupByLibrary.simpleMessage("and"),
-    "example_ExtractProminentColorsFromAnImage" : MessageLookupByLibrary.simpleMessage("Extract prominent colors from an Image"),
-    "example_ForgetPassword" : MessageLookupByLibrary.simpleMessage("Forget password?"),
-    "example_InputPassword" : MessageLookupByLibrary.simpleMessage("Please enter password"),
-    "example_InputPhoneNumber" : MessageLookupByLibrary.simpleMessage("Please enter phone number"),
-    "example_Navigator" : MessageLookupByLibrary.simpleMessage("Use Navigator for partial page jumps"),
-    "example_NewUserRegister" : MessageLookupByLibrary.simpleMessage("New User Registration"),
-    "example_PictureWaterfallFlow" : MessageLookupByLibrary.simpleMessage("Picture waterfall flow"),
-    "example_PrivacyPolicy" : MessageLookupByLibrary.simpleMessage("\"Privacy Policy\""),
-    "example_ReadAndAgree" : MessageLookupByLibrary.simpleMessage("I have read and agree"),
-    "example_UserAgreement" : MessageLookupByLibrary.simpleMessage("\"User Agreement\""),
-    "example_ValidPhoneNumberTip" : MessageLookupByLibrary.simpleMessage("Please enter a valid phone number"),
-    "home" : MessageLookupByLibrary.simpleMessage("Home"),
-    "language" : MessageLookupByLibrary.simpleMessage("Language"),
-    "login" : MessageLookupByLibrary.simpleMessage("Login"),
-    "systemDefault" : MessageLookupByLibrary.simpleMessage("System default")
-  };
+  static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
+        "account": MessageLookupByLibrary.simpleMessage("Account"),
+        "cancel": MessageLookupByLibrary.simpleMessage("Cancel"),
+        "example": MessageLookupByLibrary.simpleMessage("Example"),
+        "example_AgreeToContinueTip": MessageLookupByLibrary.simpleMessage(
+            "Need to agree to continue using"),
+        "example_And": MessageLookupByLibrary.simpleMessage("and"),
+        "example_ExtractProminentColorsFromAnImage":
+            MessageLookupByLibrary.simpleMessage(
+                "Extract prominent colors from an Image"),
+        "example_ForgetPassword":
+            MessageLookupByLibrary.simpleMessage("Forget password?"),
+        "example_InputPassword":
+            MessageLookupByLibrary.simpleMessage("Please enter password"),
+        "example_InputPhoneNumber":
+            MessageLookupByLibrary.simpleMessage("Please enter phone number"),
+        "example_Navigator": MessageLookupByLibrary.simpleMessage(
+            "Use Navigator for partial page jumps"),
+        "example_NewUserRegister":
+            MessageLookupByLibrary.simpleMessage("New User Registration"),
+        "example_PictureWaterfallFlow":
+            MessageLookupByLibrary.simpleMessage("Picture waterfall flow"),
+        "example_PrivacyPolicy":
+            MessageLookupByLibrary.simpleMessage("\"Privacy Policy\""),
+        "example_ReadAndAgree":
+            MessageLookupByLibrary.simpleMessage("I have read and agree"),
+        "example_UserAgreement":
+            MessageLookupByLibrary.simpleMessage("\"User Agreement\""),
+        "example_ValidPhoneNumberTip": MessageLookupByLibrary.simpleMessage(
+            "Please enter a valid phone number"),
+        "home": MessageLookupByLibrary.simpleMessage("Home"),
+        "language": MessageLookupByLibrary.simpleMessage("Language"),
+        "login": MessageLookupByLibrary.simpleMessage("Login"),
+        "systemDefault": MessageLookupByLibrary.simpleMessage("System default")
+      };
 }
