@@ -7,7 +7,7 @@
 // ignore_for_file:unnecessary_brace_in_string_interps, unnecessary_new
 // ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
 // ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
-// ignore_for_file:unused_import, file_names
+// ignore_for_file:unused_import, file_names, avoid_escaping_inner_quotes
 
 import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';
@@ -20,26 +20,38 @@ class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'ja';
 
   final messages = _notInlinedMessages(_notInlinedMessages);
-  static _notInlinedMessages(_) => <String, Function> {
-    "account" : MessageLookupByLibrary.simpleMessage("アカウント"),
-    "cancel" : MessageLookupByLibrary.simpleMessage("キャンセル"),
-    "example" : MessageLookupByLibrary.simpleMessage("例"),
-    "example_AgreeToContinueTip" : MessageLookupByLibrary.simpleMessage("使用を継続することに同意する必要があります"),
-    "example_And" : MessageLookupByLibrary.simpleMessage("と"),
-    "example_ExtractProminentColorsFromAnImage" : MessageLookupByLibrary.simpleMessage("画像から目立つ色を抽出す"),
-    "example_ForgetPassword" : MessageLookupByLibrary.simpleMessage("パスワードをお忘れですか？"),
-    "example_InputPassword" : MessageLookupByLibrary.simpleMessage("パスワードを入力してください"),
-    "example_InputPhoneNumber" : MessageLookupByLibrary.simpleMessage("電話番号を入力してください"),
-    "example_Navigator" : MessageLookupByLibrary.simpleMessage("使用ナビゲーター進む行局部跳転页面"),
-    "example_NewUserRegister" : MessageLookupByLibrary.simpleMessage("新しいユーザー登録"),
-    "example_PictureWaterfallFlow" : MessageLookupByLibrary.simpleMessage("画像の滝の流れ"),
-    "example_PrivacyPolicy" : MessageLookupByLibrary.simpleMessage("プライバシーポリシー "),
-    "example_ReadAndAgree" : MessageLookupByLibrary.simpleMessage("私は読んで同意しました"),
-    "example_UserAgreement" : MessageLookupByLibrary.simpleMessage("《ユーザー合意》"),
-    "example_ValidPhoneNumberTip" : MessageLookupByLibrary.simpleMessage("有効な電話番号を入力して下さい"),
-    "home" : MessageLookupByLibrary.simpleMessage("ホーム"),
-    "language" : MessageLookupByLibrary.simpleMessage("言語"),
-    "login" : MessageLookupByLibrary.simpleMessage("ログイン"),
-    "systemDefault" : MessageLookupByLibrary.simpleMessage("システムデフォルト")
-  };
+  static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
+        "account": MessageLookupByLibrary.simpleMessage("アカウント"),
+        "cancel": MessageLookupByLibrary.simpleMessage("キャンセル"),
+        "example": MessageLookupByLibrary.simpleMessage("例"),
+        "example_AgreeToContinueTip":
+            MessageLookupByLibrary.simpleMessage("使用を継続することに同意する必要があります"),
+        "example_And": MessageLookupByLibrary.simpleMessage("と"),
+        "example_ExtractProminentColorsFromAnImage":
+            MessageLookupByLibrary.simpleMessage("画像から目立つ色を抽出す"),
+        "example_ForgetPassword":
+            MessageLookupByLibrary.simpleMessage("パスワードをお忘れですか？"),
+        "example_InputPassword":
+            MessageLookupByLibrary.simpleMessage("パスワードを入力してください"),
+        "example_InputPhoneNumber":
+            MessageLookupByLibrary.simpleMessage("電話番号を入力してください"),
+        "example_Navigator":
+            MessageLookupByLibrary.simpleMessage("使用ナビゲーター進む行局部跳転页面"),
+        "example_NewUserRegister":
+            MessageLookupByLibrary.simpleMessage("新しいユーザー登録"),
+        "example_PictureWaterfallFlow":
+            MessageLookupByLibrary.simpleMessage("画像の滝の流れ"),
+        "example_PrivacyPolicy":
+            MessageLookupByLibrary.simpleMessage("プライバシーポリシー "),
+        "example_ReadAndAgree":
+            MessageLookupByLibrary.simpleMessage("私は読んで同意しました"),
+        "example_UserAgreement":
+            MessageLookupByLibrary.simpleMessage("《ユーザー合意》"),
+        "example_ValidPhoneNumberTip":
+            MessageLookupByLibrary.simpleMessage("有効な電話番号を入力して下さい"),
+        "home": MessageLookupByLibrary.simpleMessage("ホーム"),
+        "language": MessageLookupByLibrary.simpleMessage("言語"),
+        "login": MessageLookupByLibrary.simpleMessage("ログイン"),
+        "systemDefault": MessageLookupByLibrary.simpleMessage("システムデフォルト")
+      };
 }
