@@ -7,7 +7,7 @@
 // ignore_for_file:unnecessary_brace_in_string_interps, unnecessary_new
 // ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
 // ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
-// ignore_for_file:unused_import, file_names
+// ignore_for_file:unused_import, file_names, avoid_escaping_inner_quotes
 
 import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';
@@ -20,26 +20,33 @@ class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'zh_CN';
 
   final messages = _notInlinedMessages(_notInlinedMessages);
-  static _notInlinedMessages(_) => <String, Function> {
-    "account" : MessageLookupByLibrary.simpleMessage("账号"),
-    "cancel" : MessageLookupByLibrary.simpleMessage("取消"),
-    "example" : MessageLookupByLibrary.simpleMessage("示例"),
-    "example_AgreeToContinueTip" : MessageLookupByLibrary.simpleMessage("需同意才能继续使用"),
-    "example_And" : MessageLookupByLibrary.simpleMessage("和"),
-    "example_ExtractProminentColorsFromAnImage" : MessageLookupByLibrary.simpleMessage("从图像中提取突出的颜色"),
-    "example_ForgetPassword" : MessageLookupByLibrary.simpleMessage("忘记密码？"),
-    "example_InputPassword" : MessageLookupByLibrary.simpleMessage("请输入密码"),
-    "example_InputPhoneNumber" : MessageLookupByLibrary.simpleMessage("请输入手机号"),
-    "example_Navigator" : MessageLookupByLibrary.simpleMessage("使用Navigator进行局部跳转页面"),
-    "example_NewUserRegister" : MessageLookupByLibrary.simpleMessage("新用户注册"),
-    "example_PictureWaterfallFlow" : MessageLookupByLibrary.simpleMessage("图片瀑布流"),
-    "example_PrivacyPolicy" : MessageLookupByLibrary.simpleMessage("《隐私政策》"),
-    "example_ReadAndAgree" : MessageLookupByLibrary.simpleMessage("我已阅读并同意"),
-    "example_UserAgreement" : MessageLookupByLibrary.simpleMessage("《用户协议》"),
-    "example_ValidPhoneNumberTip" : MessageLookupByLibrary.simpleMessage("请输入正确的手机号"),
-    "home" : MessageLookupByLibrary.simpleMessage("首页"),
-    "language" : MessageLookupByLibrary.simpleMessage("语言"),
-    "login" : MessageLookupByLibrary.simpleMessage("登录"),
-    "systemDefault" : MessageLookupByLibrary.simpleMessage("跟随系统")
-  };
+  static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
+        "account": MessageLookupByLibrary.simpleMessage("账号"),
+        "cancel": MessageLookupByLibrary.simpleMessage("取消"),
+        "example": MessageLookupByLibrary.simpleMessage("示例"),
+        "example_AgreeToContinueTip":
+            MessageLookupByLibrary.simpleMessage("需同意才能继续使用"),
+        "example_And": MessageLookupByLibrary.simpleMessage("和"),
+        "example_ExtractProminentColorsFromAnImage":
+            MessageLookupByLibrary.simpleMessage("从图像中提取突出的颜色"),
+        "example_ForgetPassword": MessageLookupByLibrary.simpleMessage("忘记密码？"),
+        "example_InputPassword": MessageLookupByLibrary.simpleMessage("请输入密码"),
+        "example_InputPhoneNumber":
+            MessageLookupByLibrary.simpleMessage("请输入手机号"),
+        "example_Navigator":
+            MessageLookupByLibrary.simpleMessage("使用Navigator进行局部跳转页面"),
+        "example_NewUserRegister":
+            MessageLookupByLibrary.simpleMessage("新用户注册"),
+        "example_PictureWaterfallFlow":
+            MessageLookupByLibrary.simpleMessage("图片瀑布流"),
+        "example_PrivacyPolicy": MessageLookupByLibrary.simpleMessage("《隐私政策》"),
+        "example_ReadAndAgree": MessageLookupByLibrary.simpleMessage("我已阅读并同意"),
+        "example_UserAgreement": MessageLookupByLibrary.simpleMessage("《用户协议》"),
+        "example_ValidPhoneNumberTip":
+            MessageLookupByLibrary.simpleMessage("请输入正确的手机号"),
+        "home": MessageLookupByLibrary.simpleMessage("首页"),
+        "language": MessageLookupByLibrary.simpleMessage("语言"),
+        "login": MessageLookupByLibrary.simpleMessage("登录"),
+        "systemDefault": MessageLookupByLibrary.simpleMessage("跟随系统")
+      };
 }
