@@ -12,10 +12,12 @@ Effect<AccountState> buildEffect() {
 }
 
 Future<void> _onLocaleChange(Action action, Context<AccountState> ctx) async {
-  showLoading(ctx.context);
-  await onLocaleChange(action.payload);
-  ctx.broadcast(RootActionCreator.onLocaleChange());
-  ctx.broadcast(ExampleActionCreator.onLocaleChange());
-  ctx.forceUpdate();
-  dismissLoading(ctx.context);
+  if (onLocaleChange != null && action.payload != null) {
+    showLoading(ctx.context);
+    await onLocaleChange!(action.payload);
+    ctx.broadcast(RootActionCreator.onLocaleChange());
+    ctx.broadcast(ExampleActionCreator.onLocaleChange());
+    ctx.forceUpdate();
+    dismissLoading(ctx.context);
+  }
 }
