@@ -1,11 +1,13 @@
 import 'package:fish_redux/fish_redux.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_easy/flutter_easy.dart';
-import 'package:flutter_easy_example/api/tu_chong/tu_chong_model.dart';
 import 'package:flutter_easy_example/routes.dart';
 import 'package:flutter_staggered_grid_view/flutter_staggered_grid_view.dart';
 
 import 'state.dart';
+
+const int crossAxisCount = 2;
+const double spacing = 4;
 
 Widget buildView(
     TuChongTileState state, Dispatch dispatch, ViewService viewService) {
@@ -14,9 +16,9 @@ Widget buildView(
       return Column(
         children: [
           Container(
-            margin: EdgeInsets.symmetric(vertical: ImagesBean.spacing),
+            margin: EdgeInsets.symmetric(vertical: spacing),
             padding: EdgeInsets.symmetric(vertical: 10),
-            width: constraints.maxWidth - ImagesBean.spacing * 2,
+            width: constraints.maxWidth - spacing * 2,
             color: colorWithRandom(),
             child: Row(
               children: [
@@ -25,23 +27,23 @@ Widget buildView(
                   width: 50,
                   child: ClipRRect(
                       borderRadius: BorderRadius.circular(25),
-                      child: BaseWebImage(state.data.site.icon)),
+                      child: Icon(Icons.eighteen_mp)),
                 ),
                 Expanded(
                   child: Column(
                     children: [
                       BaseTitle(
-                        state.data.site.name,
+                        state.data!.title,
                         textAlign: TextAlign.center,
                       ),
                       SizedBox(width: 5),
                       BaseTitle(
-                        state.data.site.description,
+                        state.data!.siteId,
                         textAlign: TextAlign.center,
                       ),
                       SizedBox(width: 5),
                       BaseTitle(
-                        state.data.tags.join(","),
+                        state.data!.tags?.join(","),
                         textAlign: TextAlign.center,
                       ),
                     ],
@@ -51,15 +53,15 @@ Widget buildView(
             ),
           ),
           StaggeredGridView.countBuilder(
-            padding: EdgeInsets.symmetric(horizontal: ImagesBean.spacing),
+            padding: EdgeInsets.symmetric(horizontal: spacing),
             shrinkWrap: true,
             physics: NeverScrollableScrollPhysics(),
-            crossAxisCount: ImagesBean.crossAxisCount,
-            itemCount: state.data.images.length,
+            crossAxisCount: crossAxisCount,
+            itemCount: state.data?.images?.length ?? 0,
             itemBuilder: (BuildContext context, int index) => BaseButton(
               padding: EdgeInsets.zero,
               child: BaseWebImage(
-                state.data.images[index].imageURL,
+                state.data?.images?[index].imageURL,
                 placeholder: Container(
                   color: colorWithRandom(),
                 ),
@@ -71,13 +73,13 @@ Widget buildView(
               },
             ),
             staggeredTileBuilder: (int index) => new StaggeredTile.extent(
-                state.data.images[index].isSquare
-                    ? ImagesBean.crossAxisCount
+                state.data!.images![index].isSquare
+                    ? crossAxisCount
                     : 1,
-                state.data.images[index]
+                state.data!.images![index]
                     .imageHeightInWidth(constraints.maxWidth)),
-            mainAxisSpacing: ImagesBean.spacing,
-            crossAxisSpacing: ImagesBean.spacing,
+            mainAxisSpacing: spacing,
+            crossAxisSpacing: spacing,
           ),
         ],
       );
