@@ -34,21 +34,21 @@ class PackageInfoUtil {
 }
 
 /// The app name. `CFBundleDisplayName` on iOS, `application/label` on Android.
-final String? appName = ((isMacOS || isPhone) && _usePackage)
+final String appName = ((isMacOS || isPhone) && _usePackage)
     ? PackageInfoUtil.packageInfo.appName
-    : _appName;
+    : _appName ?? "";
 
 /// The app package name. `bundleIdentifier` on iOS, `getPackageName` on Android.
-final String? appPackageName = ((isMacOS || isPhone) && _usePackage)
+final String appPackageName = ((isMacOS || isPhone) && _usePackage)
     ? PackageInfoUtil.packageInfo.packageName
-    : _appPackageName;
+    : _appPackageName ?? "";
 
 /// The app package version. `CFBundleShortVersionString` on iOS, `versionName` on Android.
-final String? appVersion = ((isMacOS || isPhone) && _usePackage)
+final String appVersion = ((isMacOS || isPhone) && _usePackage)
     ? PackageInfoUtil.packageInfo.version
-    : _appVersion;
+    : _appVersion ?? "";
 
 /// The app build number. `CFBundleVersion` on iOS, `versionCode` on Android.
-final String? appBuildNumber = ((isMacOS || isPhone) && _usePackage)
+final String appBuildNumber = ((isMacOS || isPhone) && _usePackage)
     ? PackageInfoUtil.packageInfo.buildNumber
-    : _appBuildNumber;
+    : _appBuildNumber ?? "";
